NH-93561: factor db constraint checking code to a new shared module
