+++ instrumentation/jdbc/src/main/java/com/solarwinds/opentelemetry/instrumentation/TraceContextInjector.java
@@ -16,21 +16,11 @@
 
 package com.solarwinds.opentelemetry.instrumentation;
 
-import com.solarwinds.joboe.config.ConfigManager;
-import com.solarwinds.joboe.config.ConfigProperty;
 import io.opentelemetry.api.trace.Span;
 import io.opentelemetry.api.trace.SpanContext;
 import io.opentelemetry.context.Context;
-import java.util.HashSet;
-import java.util.Set;
 
 public class TraceContextInjector {
-
-  private static final Set<String> defaultDbs = new HashSet<>();
-
-  static {
-    defaultDbs.add(Db.mysql.name());
-  }
 
   public static String inject(Context context, String sql) {
     if (sql.contains("traceparent")) {
@@ -50,17 +40,5 @@ public class TraceContextInjector {
     String tag = String.format("/*traceparent='%s'*/", traceContext);
     span.setAttribute("QueryTag", tag);
     return String.format("%s %s", tag, sql);
-  }
-
-  public static boolean isDbConfigured(Db db) {
-    Set<String> configuredDbs =
-        new HashSet<>(
-            ConfigManager.getConfigOptional(ConfigProperty.AGENT_SQL_TAG_DATABASES, defaultDbs));
-    return configuredDbs.contains(db.name());
-  }
-
-  public enum Db {
-    mysql,
-    postgresql
   }
 }
